+++ src/DigitalController/5x3.cpp
@@ -21,7 +21,7 @@ namespace constants
 // Properties
 const long channel_count_default = CHANNEL_COUNT_MAX;
 
-const long power_max_default[CHANNEL_COUNT_MAX] =
+const double power_max_default[CHANNEL_COUNT_MAX] =
 {
   power_max,
   power_max,
